UCP/API: add ucp_<tag|stream>_<send|recv>_nbx
New proposed API:
 - Allow thread safety initialization of user's context per request
 - More generic and allow to unify and replace several separated APIs,
   e.g. *_nb + *_nbr = *_nbx.
 - Proof of concept doesn't show any performance degradation on
   osu_latency and osu_mbw_mr benchmarks
Functions:
 - ucp_tag_send_nbx
 - ucp_tag_recv_nbx
 - ucp_stream_send_nbx
 - ucp_stream_recv_nbx
Helper macros:
 - UCP_REQUEST_ALLOCA
 - UCP_REQUEST_CONTEXT
Types:
 - ucp_request_t
 - ucp_request_flags_t

+++ src/ucp/rma/basic_rma.c
@@ -90,7 +90,8 @@ static void ucp_rma_request_zcopy_completion(uct_completion_t *self,
 static UCS_F_ALWAYS_INLINE ucs_status_t
 ucp_rma_request_init(ucp_request_t *req, ucp_ep_h ep, const void *buffer, 
                      size_t length, uint64_t remote_addr, ucp_rkey_h rkey,
-                     uct_pending_callback_t cb, size_t zcopy_thresh, int flags)
+                     uct_pending_callback_t cb, size_t zcopy_thresh,
+                     uint64_t flags)
 {
     req->flags                = flags; /* Implicit release */
     req->send.ep              = ep;
